requirements.txt
fastapi
uvicorn
pony
psycopg2-binary
httpx
python-dotenv
